refactor: Move 'quant_module/constants.py' to the root of 'modules'
Move the constants.py file, which currently contains the paths for the
module settings directories, to the root of 'proteobenc/modules'. It
actually makes more sense to not have it in the folder of a specific
submodule because it should generally be used for all modules to store
and access the module settings directory paths.
